% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpression.R
\name{DifferentialExpression}
\alias{DifferentialExpression}
\title{Differential Expression by Inter-Type Heterogeneity}
\usage{
DifferentialExpression(CountsMatrix, Identity)
}
\arguments{
\item{CountsMatrix}{Feature x cell sparse counts matrix of class dgCMatrix}

\item{Identity}{Factor of cell identities}
}
\value{
Numeric vector of gene-wise inter-type heterogeneities
}
\description{
Differential Expression by Inter-Type Heterogeneity
}
