% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalisation.R
\name{GetFreq}
\alias{GetFreq}
\title{Normalise Frequencies}
\usage{
GetFreq(CountsMatrix)
}
\arguments{
\item{CountsMatrix}{Feature x cell sparse counts matrix of class dgCMatrix}
}
\value{
Feature x cell dense matrix of frequencies
}
\description{
Normalise Frequencies
}
