% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map.R
\name{MapFeatureSelection}
\alias{MapFeatureSelection}
\title{Feature Selection by Inter-cluster Heterogeneity}
\usage{
MapFeatureSelection(ReferenceCountsMatrix, ReferenceID, minCounts, nGenes)
}
\arguments{
\item{ReferenceCountsMatrix}{Reference count matrix; feature x cell sparse counts matrix of class dgCMatrix}

\item{ReferenceID}{Factor of reference cell identities}

\item{minCounts}{Minimum number of transcripts per gene}

\item{nGenes}{Number of genes selected}
}
\value{
Vector of top genes by inter-cluster heterogeneity
}
\description{
Feature Selection by Inter-cluster Heterogeneity
}
