% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{Population}
\alias{Population}
\title{Population Heterogeneity}
\usage{
Population(CountsMatrix)
}
\arguments{
\item{CountsMatrix}{Feature x cell sparse counts matrix of class dgCMatrix}
}
\value{
Numeric vector of gene-wise population heterogeneities
}
\description{
Population Heterogeneity
}
\examples{
set.seed(1)
CountsMatrix <- Matrix::sparseMatrix(i = sample(10, 100, replace = TRUE),
j = sample(10, 100, replace = TRUE),
x = 1L,
dims = c(10,10))
Ig <- Population(CountsMatrix)
}
