% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{Cluster}
\alias{Cluster}
\title{Single-Cell Entropic Clustering}
\usage{
Cluster(CountsMatrix, MaxClus, Multistart = 5, Seed)
}
\arguments{
\item{CountsMatrix}{Feature x cell sparse counts matrix of class dgCMatrix}

\item{MaxClus}{Maximum number of clusters}

\item{Multistart}{Number of restarts at each step}

\item{Seed}{Seed}
}
\value{
Matrix where each column is a vector of integer identities. The Nth
column encodes N clusters.
}
\description{
Single-Cell Entropic Clustering
}
