% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{FeatureSelection}
\alias{FeatureSelection}
\title{Feature Selection by Population Heterogeneity}
\usage{
FeatureSelection(CountsMatrix, minCounts = 100, nGenes = 500)
}
\arguments{
\item{CountsMatrix}{Feature x cell sparse counts matrix of class dgCMatrix}

\item{minCounts}{Minimum number of transcripts per gene}

\item{nGenes}{Number of genes selected}
}
\value{
Vector of top genes by population heterogeneity
}
\description{
Feature Selection by Population Heterogeneity
}
