% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map.R
\name{Map}
\alias{Map}
\title{Mapping of Cells Onto Reference Cell Identities}
\usage{
Map(
  MapCountsMatrix,
  ReferenceCountsMatrix,
  ReferenceID,
  minCounts = 100,
  nGenes = 500,
  Multistart = 1,
  Seed
)
}
\arguments{
\item{MapCountsMatrix}{Count matrix of cells to be mapped}

\item{ReferenceCountsMatrix}{Count matrix of cells with known identities}

\item{ReferenceID}{Factor of reference cell identities}

\item{Multistart}{Number of initial identity vectors trialed}

\item{Seed}{Seed set for both R and Python components}
}
\value{
Mapped cellular identities
}
\description{
Mapping of Cells Onto Reference Cell Identities
}
