% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{Population}
\alias{Population}
\title{Population Heterogeneity}
\usage{
Population(CountsMatrix)
}
\arguments{
\item{CountsMatrix}{Feature x cell sparse counts matrix of class dgCMatrix}
}
\value{
Numeric vector of gene-wise population heterogeneities
}
\description{
Population Heterogeneity
}
